Update matting-levin reference image to match SuiteSparse results
With commit f95dc104791e6d598d3ede6468c88be4efb28677 we fixed the build
for matting-levien to work with SuiteSparse, the successor of umfpack.

SuiteSparse produces different results for the test image in make check.
Discussing this at WilberWeek with Øyvind Kolås and Jon Nordby resulted
in the decision to update the reference image.

This might break with (older versions of) umfpack.