warp: avoid processing input when the cached output is valid
When the previously processed stroke is valid, we pass the cached
buffer directly as output, and don't need the input.  Make sure
GEGL doesn't process the op's input for nothing.

OTH, when the processed stroke is invalid, require the entire input,
since the next call to process() is going to cache it.