warp: improve abyss sampling
When a displacement vector falls outside of the input bounds, behave
as if the input had a (0, 0) vector at that point, instead of
bailing.  Note that we don't actually write any data outside the
input bounds, so even if a stroke would logically alter an out-of-
bounds pixel, it would still be read as (0, 0) by the next stroke.