sampler: make gegl_buffer_sample_cleanup() thread safe
... w.r.t. gegl_buffer_sample_foo()

Use the same mutex that guards gegl_buffer_sample(_at_level) in
gegl_buffer_sample_cleanup(), so that they can be safely called
concurrently.  This currently happens in certain ops when using
multithreading; this is probably not a very good thing to begin
with, but at least it doesn't crash anymore.