ff-load/ff-save: improve audio frame handling
Previously the audio frame handling only worked well when the audio sample-rate
was an integer multiple of the frame-rate. After these changes ff-load provides
the same number of samples for all frames, and it is the task of ff-save to
bite the correct number of samples per frame from the provided buffer.

both mp4 files and ogv files produced play back correctly in firefox, but many
media players have issues with syncing up the mp4 audio track of the mp4s,
doing some caching/code refactoring to achieve bigger compatibility would be
great.