tile-handler-cache: avoid deadlock when dropping hot tile
In gimp-tile-handler-cache.c, drop_hot_tile() might be called while
the current thread holds the global cache mutex, and subsequently
wait on the hot tile storage's mutex.  If another thread already
holds the hot tile storage's mutex, and subsequently waits on the
global cache mutex, the program deadlocks.

Fix this by temporarily unlocking the global cache mutex while
dropping the hot tile, so that if the above happens, the other
thread can progress.  This is ugly; ultimately, it would have
been nicer to order the global cache mutex lock and the hot tile
storage's mutex lock such that this couldn't happen, but we need
to lock the global cache mutex to reliably find the tile to drop
in the first place.  We do make sure, however, that unlocking the
global cache mutex in the middle of any operation that may call
drop_hot_tile() is safe.