tile-backend-swap: make thread safe; move more work to writer thread
Make the swap tile backend thread safe, by adding locks as necessary,
and by moving more of the work to the writer thread, to avoid having
to synchronize.

In particular, all reader threads share the same file destrcriptor,
so they must not try to read from the swap in parallel.  Likewise,
all manipulation of the gap list has to be synchronized, and was
moved to the writer thread.  This has the additional benefit that
storage reclamation for dead entries now happens in the writer thread,
before any outstanding write operations are performed, while, in turn,
storage allocation also happens in the writer thread, right before an
entry is written to the swap; this helps minimize the gaps between swap
entries.

Additionally, TILE_GET commands for tiles that are queued for writing
now return a COWed copy of the queued data, instead of a new copy.

This fixes ceratin artifacts in multithreaded programs.  See
bug #784907.