tile-handler-cache: don't trim still-in-use tiles
When trimming the cache, don't remove tiles whose reference count is
greater than one, and are therefore still in use.  Otherwise, if the
same tile is requested again, a different tile object might be
returned, resulting in two different tile objects referring to the
same tile.  If both tile objects are used to modify the tile data,
then the changes of only one of them are going to be comitted.

Note that, as a result, we never free hot tiles, since their reference
count is always greater than or equal to two.  This commit effectively
reverts commit 8c60f31f18ec3e91f56b4ababefef75573118478, since it's
no longer necessary.

While this problem is not multithreading-specific, it's most likely
to occur in multithreaded programs.  See bug #784907.