gcut: add support for timed global filters chains
This adds support for chains that apply on top of the cut and spliced/faded
video track. For now only chains with global time offstes preceding the clips
is supported.

Adding for instance:

-- 10.2s 20s gegl:pixelize x-size=0.1rel y-size=0.1rel

before all the clips in a video project will blur the sequence between 10.2s
and 20s.

If multiple filters apply they are applied in the order they occur in the
timeline/file.