Bug 786110 - Gaussian blur has subdivision artifacts ...
... with more than one thread

Override the new get_split_strategy() virtual function of
GeglOperationFilter in gegl:gblur-1d, so that the processed area is
split along the direction of the blur.  This avoids the chunking
artifacts when using the IIR filter, because the scanlines aren't
broken, and also improves multithreaded performance, by maximizing
the scanline length.