median-blur: change bin distribution; perform in perceptual
Change the histogram bin distribution such that, for an n-bin
histogram, the bins are centered around multiples of 1 / (n - 1).
This means that the width of the lowest and highest bins is half of
that of the rest of the bins, but also that they result in 0 and 1,
respectively, exactly.

Change the working space of the op from RGB(A) to R'G'B'(A).  Note
that, ideally, since the median depends only on the relative order
of the values, this is insignificant.  However, since we quantize
the value range into bins, this affects the distribution of the
bins, and, in particular, creates visible banding when done in RGB,
given the current bin count (even when the op in a nop, i.e., when
radius == 0).