gegl: remove big overallocation in auto-threading code
The temp-buffers are used for copies of results from
gegl_buffer_iterator in the correct format for the ops processing.
For now I have moved the allocation to be within the driving loop.

It might however be better to create an iterator per thread and cut down
on the code needed.

Fixing bug #788252.