spherize: make negative amounts perform the inverse transform
When the "amount" prop is negative, perform the inverse transform
corresponding to the positive "amount" of the same magnitude.
Previously, a negative "amount" would project the image atop the
back-face of a sunken cap, instead of the front-face of a raised
cap.  Photoshop, however, uses negative amounts to perform the
inverse transform, so, short of adding a separate "inverse" prop,
let's just roll with that.