spherize: wrap image aroudn the sphere, rather than projecting it atop it
Chnage the mapping so that the input image is wrapped around the
spherical cap, rather than projected on top of it.  In particular,
this produces an effect even when the angle of view is 0 deg.,
i.e., when the focal length is infinite.

Change the default angle of view to 0 deg.

On a side note, taking a closer look at the Photoshop Spherize
filter, it uses the "amount" property differently than us.  Both
filters produce the same effect--wrapping the image around a
hemisphere--when "amount" is 1 (and the angle of view is 0 deg.,
which can't be changed in Photoshop.)  Likewise, both filters
produce the inverse transform when "amount" is -1.  However,
while we use fractional amounts to control the curvature of the
spherical cap, by adjusting its apex angle and radius, Photoshop
uses fractional amounts to scale the full-amount displacement.
Furthermore, while we use negative amounts to perform the inverse
transform of the corresponding positive amount, Photoshop uses
negative amounts to scale the -1 displacemnt, which is *not* the
inverse transform of the corresponding positive amount.  All in
all, I like our version better, so I'm keeping it.