spherize: rename "amount" to "curvature"; add separate "amount" prop
Rename the "amount" property to "curvature" (which better conveys
its function).  Negative curvatures apply the transformation using
a concave spherical cap, having the same curvature as the convex
cap of the corresponding positive curvature (restoring the original
behavior).  Negative curvatures might be dropped later, especially
since they produce the same result as positive curvatures when the
angle of view is 0.

Add a separate "amount" property, which controls the scale factor
for the displacement (as in Photoshop).  Positive amounts use the
forward mapping, while negative amounts use the inverse mapping.