gegl/buffer: remove mipmap code from nohalo
This does the same modification that preceding commits did to lohalo, for
sigificant downsampling this might result in sharper/aliased results, and maybe
the user would be better of having chosen linear or cubic resamplers which use
a box-filter for scaling down. A similar box-filter might at a later stage be
integrated as a replacement for the mipmap using code that were in nohalo and
lohalo.

This should fix bug #763557