spherize: disable negative curvatures
Restrict the "curvature" property to the [0, 1] range.  Negative
curvatures produce the same result as positive curvatures when the
angle of view is 0, and the choice of upper bound for their mapping
to apex angles, depending on the angle of view, is arbitrary,
unlike positive curvatures.

Note that this commit only restricts the value range of the
property, while keeping support for negative curvatures in the
code, since it is trivial.