color-to-alpha-plus: add new op to workshop
color-to-alpha-plus is an extension of color-to-alpha, introducing
a pair of properties controlling the range of colors whose alpha is
modified:

  - transparency-threshold, controls the radius around the selected
    color, below which colors become fully transparent.

  - opacity-threshold, controls the radius around the selected
    color, above which colors remain fully opaque.

Colors between transparency-threshold and opacity-threshold become
partially transparent.

When transparency-threshold is 0 and opacity-threshold is 1, which
is the default configuration, the result is the same as that of
color-to-alpha.  The plan is to eventually merge the new
functionality into color-to-alpha, not to add a new op.