color-to-alpha-plus: add compress-threshold-range property
Add a boolean compress-threshold-range property.  When unset, the
transparency and opacity thresholds are interpreted as absolute
distances from the selected color (using a chessboard metric); when
set, the threshold range is compressed so that a threshold of 1 maps
to the greatest distance from the selected color to any of the faces
of the RGB cube, so that each threshold value produces a
(potentially) different result (assuming all colors are clipped).

Previously, the threshold range was always compressed, although now
we default the property to FALSE (when the transparency threshold
is 0, and the opacity threshold is 1, the result is still the same
as color-to-alpha, regardless of the value of this proeprty.)

This property is added mostly for experimentation, and will most
likely not make it back to color-to-alpha.