graph: streamline GeglVisitor
Simplify the implementation of GeglVisitor, and make it more
efficient.

Remove gegl_visitor_reset().  We don't actually use it, and the
after-traversal-but-before-reset state is pointless.  Instead, make
traversal self-contained, so there's nothing to reset.

Remove gegl_visitor_visit_visitable().  Aside from the fun name, we
don't actually use it, and it's not even defined anywhere.