graph: avoid visiting nodes multiple times during invalidation
Don't connect nodes to the INVALIDATED signal of their sources in
order to propgagte invalidation throughout the graph.  This has the
same scalability issues outlined in the previous commit.

Instead, when invalidating a node, use GeglVisitor to traverse the
output graph in breadth-first order, firing INVALIDATED signals as
we go, lumping dirty regions along the way.

This commit adds two alternative implementations for the
invalidation alogrithm.  One uses GeglRegions, which is more
accurate but has higher overhead, and the other uses
GeglRectangles, which is more coarse but has lower overhead.  We
use the GeglRectangle version for now, since it should probably be
sufficient in practice, but keep the GeglRegion version around just
in case.