gegl: do threading split of work at higher level in point ops
Slight speedup and code simplification, also removes roi miscomputations for
color and edge tests. Made possible due to increased thread safety in the
operatino of GeglBuffer. Some of this might even now be delegated to the
threading capabilities of the point base-classes parent class; for now though
the iteration moves to the threads in the leaf base-classes.