median-blur: cleanup; optimize for different formats; add "exact" prop
Some code cleanup, in particular, to remove assumptions about the
number of components, so that we can use the op natively with more
formats.

Optimize the op for different input formats, in particular, Y *,
so that we may eventually use it in GIMP to expand/contract
selections, instead of gimp:grow and gimp:shrink.

Add "exact" prop, which, when set, avoids input-value quantization
(and, in particular, clipping), and produces exact results.  This
comes at the cost of speed: performance is generally ~1.5x to ~3x
worse.

When "exact" is not set, quantize the input values into 256 bins,
instead of 1024 bins, since the higher precision is probably not
relevant to the user in this case.

Update reference hash.