median-blur: rename "exact" prop to "high-precision"
This property has no effect for R'G'B'/Y' u8 inputs, since
quantization doesn't introduce any error in these cases.  The new
name will hopefully make this fact less surprising.

In the future, it would be nice if we could use GUM to control the
visibility of properties based on the input format.