buffer: don't block while locking tiles
Allow multiple threads to lock a tile for writing concurrently,
only making sure that any required uncloning is done in a thread-
safe manner.  This is based on the assumption that the different
threads will be writing to non-overlapping regions of the tile,
which is the case for multithreaded ops, and which should be the
general use case.

As a result, unlock notifications are not guaranteed to be mutually
exclusive w.r.t. each other, or w.r.t. subsequent modifications to
the tile.