Revert "Raw load (libraw): bounding box correction for rotated images"
Other source operations that decode image file formats, eg. JPEG, do
not apply the embedded orientation. Having the RAW loader do so
complicates things for applications that use gegl:load to load pixels
from files into a GeglBuffer. If it is really necessary then it can be
done, optionally, behind a boolean property.

This is one of those things that will be addressed as part of a better
codec API for GeglBuffer. Until then, at least ensure that the loaders
consistent among themselves.

Note that due to user_flip being disabled in commit 2ca38acae33fe9db,
LibRaw wasn't actually orienting the pixels. So, the oriented bounding
box didn't correspond to the unoriented contents of the GeglBuffer.

This reverts commit 14d71e541270d0248f4307d3701c26759673205b.

https://bugzilla.gnome.org/show_bug.cgi?id=770917