operations/workshop/shadows-highlights-correction: Simplify code
It is hard to follow a maze of nested fminf and fmaxf calls. CLAMP is a
more readable way to achieve the same. However, the values are already
bounded by the limits imposed by the operation's properties. So,
calling a function to re-impose the limit needlessly complicates
things - wasteful at run-time and harder to read.

It is better to use G_LIKELY annotated assertions instead.