buffer: don't count cloned tiles towards cache_total more than once
In GeglTileHandlerCache, we currently count the size of all tiles
individually towards the cache's total size, even when some of the
tiles are clones of each other, and hence share the same data.

Improve this by counting the size of groups of mutually cloned
tiles only once, to reflect their actual memory consumption.  The
implementation uses a lock-free shared counter, similar to the way
we do COW, and has low overhead.