operation: add gegl_operation_context_dup_input_maybe_copy()
gegl_operation_context_dup_input_maybe_copy() takes a ROI, and
returns either a reference to the input buffer, or a copy of the
input buffer, containing (at least) the region required to process
ROI, if the copy has a low-enough overhead.  Note that the copied
input-buffer tiles are generally COWed, so this is rather cheap.

This is useful for threaded ops, which can use this function to
create a separate copy of the input buffer for each thread, to
avoid contention over the input's tile-storage lock.  See the next
commit.