shadows-highlights-correction: Simplify code
Since the code already scales the colour channels to [0.0, 1.0], it is
quite obvious what the maximum, half-maximum and double-maximum values
are. Moreover, variables like lmax aren't used consistently either.
eg., the code had both "1.0f - la" and "lmax - la". So, it's not clear
whether they were making it easy to understand overall algorithm, which
is otherwise undocumented.

Note that Darktable was using some of these variables to clamp the
output within [0.0, 1.0], which isn't relevant because the GEGL
operation always generates unbounded output.