perf: Test gegl:gaussian-blur's IIR filter
The infinite impulse response (IIR) filter is more interesting for
performance testing because that's what's used for most standard
deviation values, which is anything that's not too small. The current
threshold is 1.0.

By default, operations like dropshadow and shadows-highlights use the
IIR filter, and most optimization efforts in the literature are
targetted towards it. In case there is a specific need to track the
performance of the FIR filter, it can be added later.