operations/common/gblur-1d: Optimize by exploiting cache lines
Leapfrogging through the buffer for each colour channel doesn't fully
utilize the data cache. When a pixel's colour channel is accessed, it
is very likely that the other channels have also been read into the
cache because they are on the same cache line. Therefore, computing
the entire pixel at the same time avoids having to reload the same
cache line again.

https://bugzilla.gnome.org/show_bug.cgi?id=791994