shadows-highlights-correction: Optimize by avoiding NOP overlays
Overlaying the highlights or shadows layer becomes a NOP if the
corresponding xform value is 0.0. Unless the value of the "compress"
property is less than 50.0, the default being 50.0, it means that for
any given pixel only one of the layers have an effect.

Earlier it took approximately 2.5s to process an 18 megapixel JPEG with
shadows=100.0 and highlights=-100.0. Now, it takes 1.9s.