transform: fix required/invalidated rects for perspective transforms
When calculating required/invalidated rects, clip the output/input
rect to the horizon/backplane, so that none of its vertices are
past-the-horizon/behind-the-camera, and we get a proper transformed
polygon (and bounding box).

This fixes rendering artifacts caused by bad required rects,
resulting in the necessary regions of the input not being computed.