transform: don't sample past-the-horizon points
When applying a perspective transform, don't sample past-the-
horizon output points, which correspond to behind-the-camera input
points, and should be clipped.

Currently, we test each point individually, which is suboptimal.
Ideally, we should precalculate the regions that are inside the
transformed polygon, and only rasterize those.  Either way, the
output should now be the same.