transform: only rasterize inside the transformed polygon
Calculate the limits of each scanline that are inside the
transformed and clipped bounding-box of the input (with sufficient
headroom for the sampler context rect, and for box filtering), and
only rasterize inside those limits, zeroing the rest of the output.

This slightly improves performance, and avoids perspective-
transform artifacts when sampling close to the horizon.