buffer: implement bilinear dispatch along nearest
The default remains boxfilter dispatch which is our highest quality, for the
time being, bilienar dispatch for 8bit nonlinear runs at only 17% of the
performance of the much higher quality boxfilter due to lack of fast paths.
With fast paths it might be a tolerable loss of quality at some speed gain.