buffer: add loop reusing smaller tempbuf for scaled blits.
For GIMP this doesn't matter, it currently requests small enough chunks that
they are done with in one iteration of the new code. For other code that uses
GEGL to directly update a larger raster rectange, like the gegl binary or gnome
photos, the impact on cache behavior is larger. Also use stack allocations when
below GEGLs stack allocation threshold.