Bug 793937 - Crash in connection with median blur
In convert_values_to_bins(), when quantize == FALSE, sort the
values *before* initializing prev_value, since it has to initially
correspond to the value of the first bin, i.e., the first value
after sorting.  Failing to do this could lead to a buffer overflow
of bin_values, when all input values are unique.