recursive-transform: add new operation to workshop
gegl:recursive-transform applies a transformation matrix
recursilvely to an image, pasting the results on top of, or below,
each other.  This can be used to create an arbitrarily-deep image-
within-image effect.

The op additionally takes brightness/contrast parameters, which are
applied recursively as well, allowing to attenuate the colors of
deeper iterations.