buffer: improve box filtering of linear/cubic samplers
Currently, the box-filtering path of the linear/cubic samplers only
considers the diagonal of the local transformation matrix, which
works for scale matrices, but is generally wrong for arbitrary
linear transformations, as can be the case for the transform ops,
usually resulting in NaN output.

Move the common box-filtering code of the linear/cubic samplers to
gegl-sampler.h, and adjust it to take the full transformation into
account.  Namely, we take the box filtering path if the transformed
sample area is >= 4 (which, in the uniform-scale matrix case,
happens for a scale factor of 2 and above, as before), and sample
the bounding box of the transformed sample area.  This still isn't
perfect, but at least it doesn't produce invalid results.

Additionally, avoid shifting the input coordinates by 0.5px, which
isn't necessary in the box-filter case, and resulted in
inconsistencies w.r.t. the other samplers.

Update affected tests.