buffer: more improvements to sampler box filtering
In _gegl_sampler_box_get(), clamp coordinates to avoid integer
overflow, and avoid division by 0 when no samples are taken, which
can happen when the bounds of the sampled area are clamped.  In
particular, this can happen in perspective transforms where one of
the transformed vertices is close to the horizon.