buffer: fix performance regression due to last commit
The tests added in the last commit slightly worsened the
performance of sampler box filtering, even in the common case where
they aren't necessary.

Modify the code to minimize the tests and gain back the lost
performance.  While this may technically involve UB in some cases,
it should be fine in practice.