buffer: improve accuracy of sampler box filter
Previously, we sampled at uniform intervals of the bounding box of
the transformed area, which is inaccurate, as it may contain points
outside the transformed area.  Furthermore, the divergence from the
ideal result depended on the transformation, so the quality of the
result changed as the transformation changed, even when its scale
remained constant.

Instead, sample at uniform intervals along the transformed grid,
such that all sampled points are inside the transformed area.  This
is more accurate, and more consistent.

Update affected tests.

Also, Replace g_clear_pointer() with g_clear_object() in last
commit.