buffer: use linear sampler in cubic sampler box filtering
In _gegl_sampler_box_get(), make the type of sampler used for
sampling individual points configurable.  Use a NEAREST sampler for
the LINEAR sampler (i.e., when the LINEAR sampler does box
filtering, it uses a NEAREST sampler to sub-sample the image), and
a LINEAR sampler for the CUBIC sampler.  The latter is about twice
as slow, but produces better results, in particular, it is less
prone to aliasing in high-frequency input for not-too-big scale
factors.

This agrees with the general rule that CUBIC is better but slower
than LINEAR, though still notably faster than the foo-halo
samplers.