buffer: fix single-column case of gegl_buffer_{get,set}()
In the single-column case of gegl_buffer_get(), compare the output
rowstride to the output format's bpp, rather than the buffer
format's bpp, when deciding if to use the optimized path.

In the single-column case of gegl_buffer_set(), only take the
optimized path when the input rowstride equals the input format's
bpp, fix the temp-buffer's rowstride in the call to
_gegl_buffer_set_with_flags(), and return after it to avoid re-
writing to the buffer.