buffer: verify alignment in optimized buffer_set/get paths
In the optimized cases of gegl_buffer_iterate_write() and
gegl_buffer_iterate_read_simple(), make sure all read/write
accesses to the buffer and the tile are properly aligned for the
used type, and fall back to the generic version otherwise.

On some architectures, unaligned access can lead to a crash; see
bug #793682.