operations: update "gegl:value-propagate" description.
The most 2 famous filters based on this operation are "erode" and
"dilate" which are meant to work for the white and black colors.
"gegl:value-propagate" is the generic version, which works for any
color, also for alpha channel and such.
The update is mostly to add the terms "erode" and "dilate" into the
description somewhere so that this operation will be raised in the
action search in GIMP along with the basic erode/dilate filters. This
way, people will understand they can go further with it.