operations: bulk update of reference hashes
The generated files for the following operations have been visually inspected -
and the change in generated pixels are presumed to be due to other changes in
GEGL and babl. There has been subtle improvements both in color space
conversions and samplers leading to many changes. The new and updated hashes
are the hashes resulting on a 64bit linux system, running with 1 thread and no
opencl.