Bug 698468 - Add abyss-policy property to scale operations
Add an OpScale subclass of OpTransform, and use it as a base class
for the scale ops: scale-ratio, scale-size, and scale-size-
keepaspect.

OpScale provides an abyss-policy property, and overrides
OpTransform::get_abyss_policy() to return its value.  The point is
that we want to be able to control the abyss policy of scale ops
(in particular, to avoid "leaking transparency" into the image when
upscaling the entire image), while not providing this option for
other transform ops, for which it makes little sense, at least for
now.