panorama-projection: refactoring towards reverse transform
The code in the op was rigged up with the intention of doing both forward and
reverse transforms; original work on the op stopped when the panorama-preview
part worked - leaving much of the code needed for reverse in place.

This commit refactors and cleans up the reverse case, but doesn't quit make
it work yet - a bug has been filed about this remaining work in bug #795042.