operation: add GeglOperationAreaFilter::get_abyss_policy() vfunc
Add a GeglOperationAreaFilter::get_abyss_policy() virtual function,
which subclasses can override to specify which abyss policy is used
when reading the input buffer.

Modify get_required_for_output() and get_invalidated_by_change() to
take the abyss policy into account.  In particular, when using the
LOOP abyss policy, wrap the rectangle around the input bounds, so
that the correct input area is processed, and the correct output
area is invalidated.