buffer: prevent mipmapped tiles from being voided during reconstruction
While fetching a partially damaged mipmapped tile, make sure that
fetching its lower-level tiles, during the mipmapped tile's
reconstruction, doesn't cause the tile to be voided and dropped
from the cache.  This can happen if the tile handler responsible
for the lower-level tile rewrites its contents (as can happen, for
example, with a GimpTileHandlerValidate), causing the corresponding
porition of the pyramid to be voided.

We prevent the mipmapped tile from being voided by clearing its
damaged region before fetching each lower-level tile, so that any
damage to the lower-level tile can never cause the entire mipmapped
tile to be damaged.