operation: use GeglTileBackendBuffer in gegl_operation_context_dup_input_maybe_copy()
In gegl_operation_context_dup_input_maybe_copy(), use a
GeglTileBackendBuffer, backed by the input buffer, as the backend
for the local-copy buffer, instead of copying the required portion
of the input buffer into the local-copy buffer.  This allows input-
buffer tiles to be COWed on-demand, instead of upfront, into the
local-copy, which can significantly improve performance if the op's
get_required_for_output() returns a larger region than actually
necessary (in particular, if it always returns the entire input
extents).  This also allows using the local-copy optimization when
processing a level > 0.

Leave the original version #if 0-ed, since it may be useful for
testing.