noise-rgb: gummify + fix
When the op's "independent" property is FALSE, only the value of
the "red" property (and not "green" and "blue") affects the output,
using the same noise value for all channels.  Use GUM to hide the
"green" and "blue" properties in this case, and to change the label
of the "red" property to "value".

Fix output when "independent" is FALSE, and "green" or "blue" are
0 (previously, there would be no noise applied to the green/blue
channels, respectively.)