recursive-transform: merge back recursive-transform-plus changes
... which add support for applying multiple transformations
simultaneously, by specifiying a semicolon-separated list of
matrices through the "transform" property.

Note that, currently, the "fade-color" and "fade-opacity" properties
are applied to all transforms; in the future, we might want to add
per-transform fade properties.  This can be done without breaking
backward-compatibility, and the global fade properties can still be
useful.