median-blur: add "abyss-policy" property
Add an "abyss-policy" property to gegl:median-blur, which controls
the input's abyss policy, and can be one of NONE or CLAMP, for now.
When the abyss policy is NONE, the op's bounding box is expanded
according to the blur radius, while when the abyss policy is CLAMP,
the op's bounding box is clamped to the input bounding box.

The default abyss policy is CLAMP, to keep the current behavior.