buffer: fix potential race condition during tile-swap entry destruction
In gegl_tile_backend_swap_entry_destroy(), don't free the entry in
the calling thread when it has a pending op in the queue, but no
allocated storage, since the test for whether the entry has
allocated storage introduces a race condition between the calling
thread and the writer thread, where storage allocation happens.
Instead, always convert the existing op to an OP_DESTROY.

In particular, this could lead to a use-after-free and/or swap-
space leak, if a TILE_SET command is issued for a tile while it is
being written to the swap by the writer thread (after the
corresponding OP_WRITE has been dequeued).  In this case, another
OP_WRITE is pushed to the queue.  If the corresponding entry
doesn't have storage allocated yet, and is then destroyed, the
entry will be freed while the writer thread is still potentially
using it while serving the earlier OP_WRITE, and the storage
subsequently allocated for it will never be reclaimed.