buffer: always increase revision of duplicate tiles, even empty
In gegl_tile_dup(), increate the revision of the newly-created
duplicate tile unconditionally, even if it's empty, so that it gets
stored.

Instead of special-casing empty tiles in the above function, mark
newly-craeted empty tiles as stored in the empty-tile handler.
This way, newly-created empty tiles are not stored, but copied
empty tiles are, since we might need them to replace outdated data
in the tile backend.  Tile backends may, in turn, test for empty
tiles upon a SET command, and drop the existing tile instead of
storing the empty tile.

Remove the now-unnecessary revision bump in gegl_buffer_copy().